add animation on printing load
